fs: Exposed protected file/dir functions in interface class
The file/dir functions in the FileSystem interface class are a bit
complicated when dealing with access permissions. The end user shouldn't
need to access them, but other members of the FileSystem heirarchy do as
a shortcut for implementing other FileSystem classes.

The best way I can think of representing this scalably, is to make the
file/dir functions public in the FileSystem interface, but protected in
the implimentations. This does allow users to access the functions
through a cast to FileSystem, but does make it more indirect for the end
user.